DAOS-9379 java: Adjust code to adapt to new OID format (#7850)
Adjust code to adapt to DAOS API change (daos_obj_generate_oid) in v2.0
daos_obg_generate_oid API has changed in #4457. Java code needs corresponding changes.

This PR also includes "DAOS-9646 java: Turn warning into error when build JNI code".
