+++ src/client/java/daos-java/src/main/java/io/daos/DaosUtils.java
@@ -9,6 +9,7 @@ package io.daos;
 import io.daos.dfs.StatAttributes;
 
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.UUID;
@@ -137,6 +138,22 @@ public final class DaosUtils {
     return keyToBytes(key, Short.MAX_VALUE);
   }
 
+  private static byte[] padZero8(byte[] bytes) {
+    if (bytes.length >= 8) {
+      return bytes;
+    }
+
+    byte[] newBytes = new byte[8];
+    System.arraycopy(bytes, 0, newBytes, 8 - bytes.length, bytes.length);
+    Arrays.fill(newBytes, 0, 8 - bytes.length, (byte)0);
+    return newBytes;
+  }
+
+  public static byte[] keyToBytes8(String key) {
+    byte[] bytes = keyToBytes(key, Short.MAX_VALUE);
+    return padZero8(bytes);
+  }
+
   /**
    * convert key to bytes with charset {@link Constants#KEY_CHARSET}.
    *
@@ -156,5 +173,10 @@ public final class DaosUtils {
           + lenLimit);
     }
     return bytes;
+  }
+
+  public static byte[] keyToBytes8(String key, int lenLimit) {
+    byte[] bytes = keyToBytes(key, lenLimit);
+    return padZero8(bytes);
   }
 }
