+++ src/client/java/daos-java/src/main/java/io/daos/dfs/DaosFile.java
@@ -47,7 +47,7 @@ public class DaosFile {
 
   private int mode;
 
-  private DaosObjectType objectType;
+  private DaosObjectClass objectType;
 
   private int chunkSize;
 
@@ -145,7 +145,7 @@ public class DaosFile {
    * @throws IOException
    * {@link DaosIOException}
    */
-  public void createNewFile(int mode, DaosObjectType objectType, int chunkSize, boolean createParent)
+  public void createNewFile(int mode, DaosObjectClass objectType, int chunkSize, boolean createParent)
           throws IOException {
     if (objId != 0) {
       throw new IOException("file existed already");
