+++ src/client/java/daos-java/src/main/native/daos_jni_common.c
@@ -30,38 +30,32 @@ JNI_OnLoad(JavaVM *vm, void *reserved)
 			"io/daos/DaosIOException");
 
 	daos_io_exception_class = (*env)->NewGlobalRef(env, local_class);
-	jmethodID m1 = (*env)->GetMethodID(env, daos_io_exception_class,
+	new_exception_msg = (*env)->GetMethodID(env, daos_io_exception_class,
 			"<init>",
 			"(Ljava/lang/String;)V");
-
-	new_exception_msg = (*env)->NewGlobalRef(env, m1);
 	if (new_exception_msg == NULL) {
 		printf("failed to get constructor msg\n");
 		return JNI_ERR;
 	}
-	jmethodID m2 = (*env)->GetMethodID(env, daos_io_exception_class,
+	new_exception_cause = (*env)->GetMethodID(env, daos_io_exception_class,
 			"<init>",
 			"(Ljava/lang/Throwable;)V");
-
-	new_exception_cause = (*env)->NewGlobalRef(env, m2);
 	if (new_exception_cause == NULL) {
 		printf("failed to get constructor cause\n");
 		return JNI_ERR;
 	}
-	jmethodID m3 = (*env)->GetMethodID(env, daos_io_exception_class,
+	new_exception_msg_code_msg = (*env)->GetMethodID(env,
+			daos_io_exception_class,
 			"<init>",
 			"(Ljava/lang/String;ILjava/lang/String;)V");
-
-	new_exception_msg_code_msg = (*env)->NewGlobalRef(env, m3);
 	if (new_exception_msg_code_msg == NULL) {
 		printf("failed to get constructor msg, code and daos msg\n");
 		return JNI_ERR;
 	}
-	jmethodID m4 = (*env)->GetMethodID(env, daos_io_exception_class,
+	new_exception_msg_code_cause = (*env)->GetMethodID(env,
+			daos_io_exception_class,
 			"<init>",
 			"(Ljava/lang/String;ILjava/lang/Throwable;)V");
-
-	new_exception_msg_code_cause = (*env)->NewGlobalRef(env, m4);
 	if (new_exception_msg_code_cause == NULL) {
 		printf("failed to get constructor msg, code and cause\n");
 		return JNI_ERR;
@@ -71,11 +65,6 @@ JNI_OnLoad(JavaVM *vm, void *reserved)
 	if (rc) {
 		printf("daos_init() failed with rc = %d\n", rc);
 		printf("error msg: %.256s\n", d_errstr(rc));
-		return rc;
-	}
-	rc = daos_eq_lib_init();
-	if (rc) {
-		printf("Failed daos_eq_lib_init: %d\n", rc);
 		return rc;
 	}
 	return JNI_VERSION;
@@ -92,7 +81,7 @@ throw_base(JNIEnv *env, char *msg, int error_code,
 		const char *temp = posix_error ?
 				strerror(error_code) : d_errstr(error_code);
 
-		daos_msg = temp;
+		daos_msg = (char *)temp;
 	} else {
 		daos_msg = NULL;
 	}
@@ -146,10 +135,5 @@ JNI_OnUnload(JavaVM *vm, void *reserved)
 		return;
 	}
 	(*env)->DeleteGlobalRef(env, daos_io_exception_class);
-	(*env)->DeleteGlobalRef(env, new_exception_msg);
-	(*env)->DeleteGlobalRef(env, new_exception_cause);
-	(*env)->DeleteGlobalRef(env, new_exception_msg_code_msg);
-	(*env)->DeleteGlobalRef(env, new_exception_msg_code_cause);
-	daos_eq_lib_fini();
 	daos_fini();
 }
