+++ src/client/java/daos-java/src/main/native/io_daos_obj_DaosObjClient.c
@@ -4,6 +4,8 @@
  * SPDX-License-Identifier: BSD-2-Clause-Patent
  */
 
+#define _GNU_SOURCE
+
 #include "io_daos_dfs_DaosFsClient.h"
 #include "DaosObjectAttribute.pb-c.h"
 #include <daos.h>
@@ -20,30 +22,24 @@ parse_object_id(char *buffer, daos_obj_id_t *oid)
 
 JNIEXPORT void JNICALL
 Java_io_daos_obj_DaosObjClient_encodeObjectId(JNIEnv *env, jclass clientClass,
-		jlong oidBufferAddress, jint feats,
-		jstring objectClass, jint args)
+		jlong oidBufferAddress, jlong contHandle, jint objectType,
+		jstring objectClass, jint hint, jint args)
 {
 	daos_obj_id_t oid;
-	uint16_t type;
+	daos_handle_t coh;
+	enum daos_otype_t otype = (enum daos_otype_t)objectType;
+	uint32_t oclass;
 	const char *oclass_name = (*env)->GetStringUTFChars(env, objectClass,
 			NULL);
 	char *buffer = (char *)oidBufferAddress;
 
-	type = daos_oclass_name2id(oclass_name);
-	if (!type) {
-		char *tmp = "unsupported object class, %s";
-		char *msg = NULL;
-
-		asprintf(&msg, tmp, oclass_name);
-		throw_obj(env, msg, CUSTOM_ERR6);
-		goto out;
-	}
+	memcpy(&coh, &contHandle, sizeof(coh));
+	oclass = daos_oclass_name2id(oclass_name);
 	parse_object_id(buffer, &oid);
-	daos_obj_generate_id(&oid, feats, type, args);
+	daos_obj_generate_oid(coh, &oid, otype, oclass, hint, args);
 	memcpy(buffer, &oid.hi, 8);
 	memcpy(buffer + 8, &oid.lo, 8);
 
-out:
 	(*env)->ReleaseStringUTFChars(env, objectClass, oclass_name);
 }
 
@@ -65,7 +61,7 @@ Java_io_daos_obj_DaosObjClient_openObject(JNIEnv *env, jclass clientClass,
 		char *tmp = "Failed to open DAOS object with mode (%d)";
 		char *msg = NULL;
 
-		asprintf(msg, tmp, mode);
+		asprintf(&msg, tmp, mode);
 		throw_obj(env, msg, rc);
 		return -1;
 	}
@@ -288,13 +284,13 @@ decode_initial(data_desc_t *desc, char *desc_buffer)
 		/* sgl */
 		memcpy(&value64, desc_buffer, 8);
 		desc_buffer += 8;
-		d_iov_set(&desc->iovs[i], value64,
+		d_iov_set(&desc->iovs[i], (void *)value64,
 			nbr_of_record * desc->record_size);
 		desc->sgls[i].sg_iovs = &desc->iovs[i];
 		desc->sgls[i].sg_nr = 1;
 		desc->sgls[i].sg_nr_out = 0;
 	}
-	desc->ret_buf_address = desc_buffer;
+	desc->ret_buf_address = (uint64_t)desc_buffer;
 	return 0;
 }
 
@@ -482,11 +478,11 @@ decode(JNIEnv *env, jlong objectHandle, jint nbrOfAkeys,
 	desc_buffer += 2;
 	d_iov_set(dkey, desc_buffer, len);
 	desc_buffer += len;
-	if (desc_buffer - descBufAddress > descBufCap) {
+	if ((uint64_t)desc_buffer - descBufAddress > descBufCap) {
 		char *msg = NULL;
 
 		asprintf(&msg, "length %ld exceeds buffer capacity %d",
-			 desc_buffer - descBufAddress, descBufCap);
+			 (uint64_t)desc_buffer - descBufAddress, descBufCap);
 		throw_obj(env, msg, CUSTOM_ERR7);
 		return 1;
 	}
@@ -767,12 +763,12 @@ allocate_simple_desc(char *descBufAddress, data_desc_simple_t *desc,
 		/* sgl */
 		memcpy(&address, desc_buffer, 8);
 		desc_buffer += 8;
-		d_iov_set(&desc->iovs[i], address, 0);
+		d_iov_set(&desc->iovs[i], (void *)address, 0);
 		desc->sgls[i].sg_iovs = &desc->iovs[i];
 		desc->sgls[i].sg_nr = 1;
 		desc->sgls[i].sg_nr_out = 0;
 	}
-	desc->ret_buf_address = desc_buffer;
+	desc->ret_buf_address = (uint64_t)desc_buffer;
 	/* put address to the start of desc buffer */
 	memcpy((char *)descBufAddress, &desc, 8);
 	return 0;
@@ -829,7 +825,7 @@ allocate_desc_update_async(char *descBuf, data_desc_upd_async_t *desc,
 	desc->sgls[0].sg_iovs = &desc->iovs[0];
 	desc->sgls[0].sg_nr = 1;
 	desc->sgls[0].sg_nr_out = 0;
-	desc->ret_buf_address = desc_buffer;
+	desc->ret_buf_address = (uint64_t)desc_buffer;
 	/* put address to the start of desc buffer */
 	if (reuse) {
 		memcpy(descBuf - 8, &desc, 8);
@@ -969,7 +965,7 @@ static int
 update_ret_code(void *udata, daos_event_t *ev, int ret)
 {
 	data_desc_simple_t *desc = (data_desc_simple_t *)udata;
-	char *desc_buffer = desc->ret_buf_address;
+	char *desc_buffer = (char *)desc->ret_buf_address;
 
 	memcpy(desc_buffer, &ret, 4);
 	if (ev) {
@@ -1019,7 +1015,7 @@ static int
 update_ret_code_no_decode(void *udata, daos_event_t *ev, int ret)
 {
 	data_desc_upd_async_t *desc = (data_desc_upd_async_t *)udata;
-	char *desc_buffer = desc->ret_buf_address;
+	char *desc_buffer = (char *)desc->ret_buf_address;
 
 	memcpy(desc_buffer, &ret, 4);
 	if (ev) {
@@ -1215,12 +1211,12 @@ decode_async(JNIEnv *env, jlong descBufAddress,
 		/* sgl */
 		memcpy(&value64, desc_buffer, 8);
 		desc_buffer += 8;
-		d_iov_set(&desc->iovs[i], value64, value);
+		d_iov_set(&desc->iovs[i], (void *)value64, value);
 		desc->sgls[i].sg_iovs = &desc->iovs[i];
 		desc->sgls[i].sg_nr = 1;
 		desc->sgls[i].sg_nr_out = 0;
 	}
-	desc->ret_buf_address = desc_buffer;
+	desc->ret_buf_address = (uint64_t)desc_buffer;
 	*ret_desc = desc;
 	return 0;
 }
@@ -1229,7 +1225,7 @@ static int
 update_ret_code_async(void *udata, daos_event_t *ev, int ret)
 {
 	data_desc_async_t *desc = (data_desc_async_t *)udata;
-	char *desc_buffer = desc->ret_buf_address;
+	char *desc_buffer = (char *)desc->ret_buf_address;
 
 	memcpy(desc_buffer, &ret, 4);
 	desc->event->status = 0;
@@ -1252,7 +1248,7 @@ Java_io_daos_obj_DaosObjClient_updateObjectAsync(
 	if (rc) {
 		char *msg = "Failed to allocate memory";
 
-		throw_exception_const_msg_object(env, msg, rc);
+		throw_const_obj(env, msg, rc);
 		goto fail;
 	}
 	desc->event->status = EVENT_IN_USE;
@@ -1262,7 +1258,7 @@ Java_io_daos_obj_DaosObjClient_updateObjectAsync(
 	if (rc) {
 		char *msg = "Failed to register update callback";
 
-		throw_exception_const_msg_object(env, msg, rc);
+		throw_const_obj(env, msg, rc);
 		goto fail;
 	}
 	rc = daos_obj_update(oh, DAOS_TX_NONE, flags, &desc->dkey,
@@ -1271,7 +1267,7 @@ Java_io_daos_obj_DaosObjClient_updateObjectAsync(
 	if (rc) {
 		char *msg = "Failed to update DAOS object asynchronously";
 
-		throw_exception_const_msg_object(env, msg, rc);
+		throw_const_obj(env, msg, rc);
 		goto fail;
 	}
 	return;
@@ -1284,7 +1280,7 @@ static int
 update_actual_size(void *udata, daos_event_t *ev, int ret)
 {
 	data_desc_simple_t *desc = (data_desc_simple_t *)udata;
-	char *desc_buffer = desc->ret_buf_address;
+	char *desc_buffer = (char *)desc->ret_buf_address;
 	int i;
 	uint32_t value;
 
@@ -1350,7 +1346,7 @@ static int
 update_actual_size_async(void *udata, daos_event_t *ev, int ret)
 {
 	data_desc_async_t *desc = (data_desc_async_t *)udata;
-	char *desc_buffer = desc->ret_buf_address;
+	char *desc_buffer = (char *)desc->ret_buf_address;
 	int i;
 	uint32_t value;
 
@@ -1382,7 +1378,7 @@ Java_io_daos_obj_DaosObjClient_fetchObjectAsync(
 	if (rc) {
 		char *msg = "Failed to allocate memory";
 
-		throw_exception_const_msg_object(env, msg, rc);
+		throw_const_obj(env, msg, rc);
 		goto fail;
 	}
 	desc->event->status = EVENT_IN_USE;
@@ -1392,7 +1388,7 @@ Java_io_daos_obj_DaosObjClient_fetchObjectAsync(
 	if (rc) {
 		char *msg = "Failed to register fetch callback";
 
-		throw_exception_const_msg_object(env, msg, rc);
+		throw_const_obj(env, msg, rc);
 		goto fail;
 	}
 	rc = daos_obj_fetch(oh, DAOS_TX_NONE, flags, &desc->dkey,
@@ -1401,7 +1397,7 @@ Java_io_daos_obj_DaosObjClient_fetchObjectAsync(
 	if (rc) {
 		char *msg = "Failed to fetch DAOS object asynchronously";
 
-		throw_exception_const_msg_object(env, msg, rc);
+		throw_const_obj(env, msg, rc);
 		goto fail;
 	}
 	return;
